refactor: move all voting escrow tests to single file
